% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket_mean.R
\name{bracket_mean}
\alias{bracket_mean}
\title{Bracket-mean approximation}
\usage{
bracket_mean(n = 3, p = rep(1/n, n), q_fun, d_fun, params)
}
\arguments{
\item{n}{numeric of length 1; represents a number of intervals (or a number
of point of the approximation). Default value is 3. Either \code{n} or
\code{p} should be used.}

\item{p}{numeric of length grater than 2; specifies the corresponding
probabilities of intervals. If \code{p} does not sum up to 1, will be
automatically standardized.}

\item{q_fun}{quantile function.}

\item{d_fun}{corresponding probability density function, which will be used
for integral estimation.}

\item{params}{a list of parameters of the distribution.}
}
\value{
a list of pairs probability-outcome.
}
\description{
Returns a discrete approximation of a continuous distribution using
bracket-mean method.
}
\details{
If \code{n} is specified, the function splits the range of possible values
into \code{n} intervals by using 1/n, 2/n, ... (n-1)/n-quantiles. Otherwise,
if \code{p} is specified, the range is split by corresponding cumulative
probabilities quantiles. \code{q_fun} is used to calculate quantiles. Each
interval is represented by pair of probability and a mean of this interval,
which calculated as an integral over this interval.
}
\examples{
bracket_mean(n = 5, q_fun = qnorm, d_fun = dnorm,
             params = list(mean = 3, sd = 2))
bracket_mean(p = c(1/2, 1/3, 1/6), q_fun = qexp, d_fun = dexp,
             params = list(rate = 1))
bracket_mean(q_fun = qnorm, d_fun = dnorm)

}
\references{
\itemize{
\item Miller III A.C., Rice T.R. \emph{Discrete Approximations of Probability
Distributions}. Management Science, 29(3):352-362, 1983.
\item Smith J.E. \emph{Moment Methods for Decision Analysis}. Management
Science 39(3):340-358, 1993.
}
}
\seealso{
\code{\link{bracket_median}}
}
