% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket_median.R
\name{bracket_median}
\alias{bracket_median}
\title{Bracket-median approximation}
\usage{
bracket_median(n = 3, p = rep(1/n, n), q_fun, params)
}
\arguments{
\item{n}{numeric of length 1; represents a number of intervals (or a number
of point of the approximation). Default value is 3. Either \code{n} or
\code{p} should be used.}

\item{p}{numeric of length grater than 2; specifies the corresponding
probabilities of intervals. If \code{p} does not sum up to 1, will be
automatically standardized.}

\item{q_fun}{quantile function.}

\item{params}{a list of parameters of the distribution.}
}
\value{
a list of pairs probability-outcome.
}
\description{
Returns a discrete approximation of a continuous distribution using
bracket-median method.
}
\details{
If \code{n} is specified, the function splits the range of possible values
into \code{n} intervals by using 1/n, 2/n, ... (n-1)/n-quantiles. Otherwise,
if \code{p} is specified, the range is split by corresponding cumulative
probabilities quantiles. \code{q_fun} is used to calculate quantiles. Each
interval is represented by pair of probability and median of the interval.
}
\examples{
bracket_median(n = 5, q_fun = qnorm, params = list(mean = 3, sd = 2))
bracket_median(p = c(1/2, 1/3, 1/6), q_fun = qexp, params = list(rate = 1))
bracket_median(q_fun = qnorm)
# using sample data
sample <- rnorm(100) + rexp(100)
q <- function(p) quantile(x = sample, probs = p)
bracket_median(q_fun = q)

}
\references{
\itemize{
\item Miller III A.C., Rice T.R. \emph{Discrete Approximations of Probability
Distributions}. Management Science, 29(3):352-362, 1983.
\item Smith J.E. \emph{Moment Methods for Decision Analysis}. Management
Science 39(3):340-358, 1993.
\item Keefer D.L., Bodily S.E. \emph{Three-Point Approximations for
Continuous Random Variables}. Management Science 29(5):595-609, 1983.
}
}
\seealso{
\code{\link{bracket_mean}}
}
