% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_swanson_megill.R
\name{extended_swanson_megill}
\alias{extended_swanson_megill}
\title{Extended Swanson-Megill Three-Point approximation}
\usage{
extended_swanson_megill(q_fun, params)
}
\arguments{
\item{q_fun}{quantile function.}

\item{params}{a list of parameters of the distribution.}
}
\value{
a list of pairs probability-outcome.
}
\description{
Returns a discrete approximation of a continuous distribution using
extended Swanson-Megill method.
}
\details{
Method approximate a continuous distribution by 0.1, 0.5, and 0.9 quantiles
with probabilities 0.3, 0.4, and 0.3, respectively.
}
\examples{
extended_swanson_megill(q_fun = qnorm, params = list(mean = 3, sd = 2))
extended_swanson_megill(q_fun = qgamma, params = list(shape = 3, rate = 2.2))
extended_swanson_megill(q_fun = qexp)
# using sample data
sample <- rnorm(100) + rexp(100)
q <- function(p) quantile(x = sample, probs = p)
extended_swanson_megill(q_fun = q)

}
\references{
\itemize{
\item Keefer D.L., Bodily S.E. \emph{Three-Point Approximations for
Continuous Random Variables}. Management Science 29(5):595-609, 1983.
\item Keefer L.D. \emph{Certainty Equivalents for Three-Point Discrete-
Distribution Approximations}, Management Science 40(6): 760-773, 1994.
}
}
\seealso{
\code{\link{extended_pearson_tukey}}, \code{\link{mcnamee_celona}}
}

