% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnamee_celona.R
\name{mcnamee_celona}
\alias{mcnamee_celona}
\title{McNamee-Celona "Shortcut" approximation}
\usage{
mcnamee_celona(q_fun, params)
}
\arguments{
\item{q_fun}{quantile function.}

\item{params}{a list of parameters of the distribution.}
}
\value{
a list of pairs probability-outcome.
}
\description{
Returns a discrete approximation of a continuous distribution using
McNamee-Celona "Shortcut" method.
}
\details{
Method approximate a continuous distribution by 0.1, 0.5, and 0.9 quantiles
with probabilities 0.25, 0.5, and 0.25, respectively.
}
\examples{
mcnamee_celona(q_fun = qnorm, params = list(mean = 3, sd = 2))
mcnamee_celona(q_fun = qgamma, params = list(shape = 3, rate = 2.2))
mcnamee_celona(q_fun = qexp)
# using sample data
sample <- rnorm(100) + rexp(100)
q <- function(p) quantile(x = sample, probs = p)
mcnamee_celona(q_fun = q)

}
\references{
\itemize{
\item McNamee P., Celona J. \emph{Decision Analysis for the Professional
with Supertree}, Scientific Press, Redwood City, CA, 1987.
\item Keefer L.D. \emph{Certainty Equivalents for Three-Point Discrete-
Distribution Approximations}, Management Science 40(6): 760-773, 1994.
}
}
\seealso{
\code{\link{extended_swanson_megill}},
\code{\link{extended_pearson_tukey}}, \code{\link{brown_kahr_peterson_3}}
}

