% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown_kahr_peterson_3.R
\name{brown_kahr_peterson_3}
\alias{brown_kahr_peterson_3}
\title{Three-point approximation, proposed by Brown, Kahr, and Peterson}
\usage{
brown_kahr_peterson_3(q_fun, params)
}
\arguments{
\item{q_fun}{quantile function.}

\item{params}{a list of parameters of the distribution.}
}
\value{
a list of pairs probability-outcome.
}
\description{
Returns a discrete approximation of a continuous distribution using
three-point approximation, proposed by Brown, Kahr, and Peterson.
}
\details{
Note: only distribution with finite quantiles are supported.
}
\examples{
brown_kahr_peterson_3(q_fun = qbeta, params = list(shape1 = 2, shape2 = 3))
brown_kahr_peterson_3(q_fun = qunif)
# using sample data
sample <- rnorm(100) + rexp(100)
q <- function(p) quantile(x = sample, probs = p)
brown_kahr_peterson_3(q_fun = q)

\dontrun{
brown_kahr_peterson_3(q_fun = qnorm)
}

}
\references{
\itemize{
\item Smith J.E. \emph{Moment Methods for Decision Analysis}. Management
Science 39(3):340-358, 1993.
\item Keefer D.L., Bodily S.E. \emph{Three-Point Approximations for
Continuous Random Variables}. Management Science 29(5):595-609, 1983.
\item Pearson E.S., Tukey J.W. \emph{Approximate Means and Standard Deviations
Based on Distances between Percentage Points of Frequency Curves}.
Biometrika 52(3/4):533-546, 1965.
}
}
\seealso{
\code{\link{extended_swanson_megill}}, \code{\link{mcnamee_celona}}
}

